U430-029 fix assertion to prevent miscompilation
On test Q802-001__stringmap continuous builder was reporting a
Storage_Error which was not present in developer builds.

Running gnat2why on this test under valgrind reported an access to
uninitialized memory when accessing a non-existing variant record
component. This looks like a GNAT bug, because we should be getting
a run-time check, but since the gnat2why code was wrong, we need to
fix it anyway.

* src/flow/flow-control_flow_graph.adb
(Do_Object_Declaration): Add guard before accessing the F.Component.

Change-Id: Ib78fd32daddf849a634a4e0bc2f7dd5361b80155