U218-042 sync generation of Initializes in phases 1 and 2
When generating Initializes in phase 1, we computed intersection of pure
outputs with local variables of a package as they are seen in the
package body, but in phase 2 we computed intersection with local
variables as they are seen in the package spec. Now in phase 2 we do
exactly what we did in phase 1.

This requires us to record local variables in the ALI file as they are
seen in the body. We then up-project them when combining variables
contributed by a child package with those owned by the parent package.

This change causes no diffs (as expected), except for a degenerated test
with missing Part_Of indicators, where objects declared in the private
part of a package are implicitly lifted into singleton abstract states
and appear in the generated Initialized contracts. Previously, they were
filtered by the intersection with objects that can legally appear on the
LHS of the Initializes clauses.

* src/flow/flow-control_flow_graph.adb
(Register_Own_Variable): We need local variables both when Initializes
needs to be generated and when those variables might be written by a
parent unit.
* src/flow/flow_generated_globals-partial.adb
(Preanalyze_Body): Pick local variables even if the Initializes contract
is not generated; the exact condition is in Register_Own_Variable.
(Write_Contracts_To_ALI): Use local variables as they are seen in the
package body, not in the package spec.
* src/flow/flow_generated_globals-phase_2.adb
(Fold): Do exactly what we do in phase 1.

* testsuite/gnatprove/tests/O909-059__part_of/test.out
New messages are not really better, but the generated Initializes
contract actually makes more sense. It is just that we don't have
enough protection from cascaded errors related to missing Part_Ofs.

Change-Id: I3c11068f0852f39d8738cfb6ac930d80b9d63e7b