PA20-030 More use of precise AST types with predicates
Now applied to subprogram utilities.

Change the Subprogram_Id subtype to better reflect which entities are
treated as subprograms in GNATprove, by adding E_Subprogram_Type for
indirect calls through access to subprogram. More generally, we should
rethink subtypes here, as sometimes tasks are also treated like
subprograms. It would be an improvement to identify suitable subtypes
that would be used more consistently.

[changelog]
* src/flow/flow_generated_globals-partial.adb (Preanalyze_Spec): Add
guard to call to Is_Predefined_Potentially_Blocking.
* src/flow/flow_refinement.adb: Reformat following type renaming.
* src/flow/flow_refinement.ads: Reformat following type renaming.
* src/spark/spark_util-subprograms.adb: Retype.
* src/spark/spark_util-subprograms.ads: Retype.
* src/spark/spark_util.adb: Rename type.
* src/spark/spark_util.ads: Rename type.
* src/utils/checked_types.ads: Introduce necessary subtypes. Rename
subtype Subprogram_Id to Subprogram_Like_Kind_Id, to be more inline with
current naming conventions. Include E_Subprogram_Type in this subtype,
as this is also treated as a subprogram in GNATprove.

Change-Id: Ib0532a7c7e242d3c450fcaf9aec45fa9732959a4