U709-012 remove warning on unused global with null dependency
This patch suppresses the warning message "... is set by ... but not
used after the call" on global parameters when a Depends contract
"... => null" is present. It reflects the behavior adopted in T117-040
for regular parameters.

* src/flow/flow-analysis.adb
(Null_Dependency_Assignment): Allow to call on a global parameter.
(Find_Ineffective_Statements): Do not raise the warning about a
  global parameter when a Depends contract "... => null" is present.

* testsuite/gnatprove/tests/M219-040__flow_IFA_error_messages/test.out
* testsuite/gnatprove/tests/RM_Examples/test.out
* testsuite/gnatprove/tests/TA06-037__depends_child/test.out
* .../tests/U709-012__flow_unused_global_null_dependency/test.out
Update tests outputs.

Change-Id: Ib5da7ed95acad6c629aecf303bb2f9c156ad6743