U715-015 Consistent handling of predefined units
Define a common proxy to decide when a node should be ignored because
it is defined in an internal (predefined or GNAT-specific) unit and
the main unit being analyzed is *not* an internal unit. Use that query
consistently across gnat2why code.

Lib.In_Predefined_Unit is only called once in spark_util.adb to define
Is_Predefined_Initialized_Entity, and once in flow-analysis.adb for
specializing an error message. Lib.In_Internal_Unit is only called
to define the new query.

[changelog]
* src/flow/flow-analysis.adb: Keep using query on predefined unit,
as this is not a question here of ignoring the node, but on issuing
an appropriate error message.
* src/flow/flow.adb: Same.
* src/flow/flow_error_messages.adb: Same.
* src/flow/flow_generated_globals-partial.adb: Same.
* src/flow/flow_generated_globals-phase_2-visibility.adb: Same.
* src/flow/flow_generated_globals-phase_2.ads: Same.
* src/flow/flow_sanity.adb: Same.
* src/flow/flow_utility.adb: Same.
* src/spark/spark_definition.adb: Same.
* src/spark/spark_util.ads (Is_Ignored_Internal): Define the query.
* src/why/gnat2why-util.adb: Use new query.

Change-Id: Id98bd1112138fb4462c7283c85f0d67cc7cf36b1