U618-036 fix antialiasing internals to address sanity check
Fix the internal behaviour of antialising to address concerns
revealed by sanity test code.
When generating the "path to root", if a node appears root-like
then check to see if it is actually a rewritten node. If so,
then check the original version - if that is "interesting" and
not the root, use the original version and continure the
excavation.

* src/flow/flow-analysis-antialiasing.adb
(Path_To_Root) One-line fix which effectviely implements the clause
"use the original version" by inserting the original version of the
node into the path to root.

Change-Id: If16cb89032cf5855c83fc4e3f5ddad985e3a93de