U715-020 simplify handling of discriminants
In SPARK, we don't support renaming of discriminants. We can simplify
the handling in proof with that assumption (but not in marking).

* src/spark/spark_util.adb
(Is_Not_Hidden_Discriminant): Do not check for completely hidden
 discriminants, they cannot occur in SPARK code.
(Root_Discriminant): As discriminants cannot be renamed, it is enough to
 search the component by name in the root of the class if it has
 discriminants.

* src/why/spark_atree-entities.adb
(First_Discriminant, Next_Discriminant, Has_Discriminant): In a type,
 discriminants are either all hidden or all visible. Simplify the code.

* tests/U715-020__hidden_discriminant
File test.

Change-Id: I66ce048ef3ac0b1cc2c5736e94885dbfa87c514a