TC02-027 Test SPARK_RAC in testsuite
Add command line option --debug-exec-rac to gnatprove (and
gnat2why) to only execute the main procedure of the project
using RAC, and terminate. This option is used to test the
RAC in the testsuite by comparing the output of the compiled
executable and the RAC.

* src/common/gnat2why_opts.ads:
* src/common/gnat2why_opts.adb:
* src/gnatprove/configuration.ads:
* src/gnatprove/configuration.adb:
* src/spark/gnat2why_opts-reading.adb:
* src/spark/gnat2why_opts-reading.ads:
* src/spark/gnat2why_opts-reading.adb:
Add option --debug-exec-rac to gnatprove and pass to gnat2why as
debug_exec_rac.

* src/why/gnat2why-driver.adb:
Execute RAC on main procedure when --debug-exec-rac=on.

* share/spark/help.txt
Document command line option --debug-exec-rac

* testsuite/gnatprove/run-tests:
Add test driver for tests called rac__*, which compiles an
executable and compares the output with the output of gnatprove
--debug-exec-rac=on.

* testsuite/gnatprove/tests/rac__test/:
Add tests for RAC.

Change-Id: Ib10fb6a1bb78bf1dfb252e18d55b5c8f0ef7e0c0