U715-021 fix location of predicate check
For assertions of the form "Expr in Type_With_Pred", if the
predicate had several conjunctions, an unproved assertion was
misplaced at the location of the unproved part of the predicate.

We use the existing GP_Inlined marker to avoid using the location of
inlined expressions. This marker was previously only generated
inside gnatwhy3.

* vc_kinds.ads
  add new marker constant
* gnat2why-expr.adb
(Compute_Dynamic_Predicate): Add GP_Inlined label to the inlined
  predicate expression
* U715-021__predicate: new test

Change-Id: I1942f6720bf214a76a7194408e84e1c121a600b3