TA21-030 suppress spurious codepeer violation
This is a temporary solution to the spurious message that codepeer
emits:

  high warning: subp always fails:
  report_database.pragma_assume'Initialize fails for all possible inputs

Unfortunately, there is currently no way to suppress a message of a kind
"subp always fails" inside codepeer, so this commit does the suppression
within the GPR file.

U804-011 has been raised with the codepeer team to fix this inside
codepeer.

Change-Id: Ia5f1acd9cc3b0278c4a2a853b88614b349ca2f8f