More precise handling of Unchecked_Deallocation in borrow checker
The argument of a call to an instance of Ada.Unchecked_Deallocation
might be partially moved, which was carried over to the permission of
the argument after the call. Instead, the argument has permission RW
after the call (and a value of null).

[changelog]
* src/why/gnat2why-borrow_checker.adb (Process_Path): Handle case of
parameter to Unchecked_Deallocation instance.
* testsuite/gnatprove/tests/U824-033__free/test.adb
* testsuite/gnatprove/tests/U824-033__free/test.out
Add test.

Change-Id: I8e66f15600bf2930c73392793d5b941da75e093d
TN: U824-033