U810-002 document expressions of type root_real as not in SPARK
Expressions of type root_real are either statically evaluated by the
frontend or rejected in marking (which was done in QB19-001 but never
documented).

Apparently we only get such expressions when using the predefined
exponentiation operator (**), but it is better to describe them in
general as not allowed in SPARK.

* docs/lrm/source/declarations-and-types.rst
(Scalar types): Add a subsubsection about real types.

Change-Id: I48312075b6f25c9ab66c9ababa07ed3931658492