U907-022 guard post axiom of dispatching functions with equality of tags
The axiom generated for the classwide postconditions of primitives of
tagged types should be guarded by a predicate ensuring that the tags of
the controlling parameters if any are equal to the tag controlling the
call.

* src/why/gnat2why-subprograms.adb
(Compute_Guard_For_Dispatch): Compute a predicate stating that the tags
 of the controlling parameters of a subprogram if any are equal to the
 tag controlling the call.
(Generate_Axiom_For_Post): Use the new predicate to guard the
 postcondition of dispatching function calls.

* tests/U907-022__dispatching_post
File test.

* tests/PB07-039__bad_mlw_with_flow_and_proof/test.out
Spurious tag check becomes unproved. It should be removed, see U908-014.

Change-Id: I7d5c79b69bb4aa80b5d624d8f3d54722a30a8698