U908-014 add tag update and conversion on inherited function calls
* src/spark/spark_util-types.adb
(Is_Derived_Type_With_Null_Ext): Check is a type is declared as a
 derived type with a null record extension. It is only used in ghost
 code currently.

* src/why/gnat2why-expr.adb
(Transform_Function_Call): If the function has a controlling result, the
 call might not have the correct tag if the function was inherited.
 Update the tag explicitly after the call and introduce checks if
 necessary.

* src/why/why-gen-records.adb
(New_Tag_Update): Do not update the type of the expression, only the
 tag.

* src/why/gnat2why-subprograms.adb
(Generate_Dispatch_Compatibility_Axioms): In the compatibility axiom of
 inherited function with controlling result, update the tag of the
 specific function call.

* tests/U908-014__inherited_function_calls
File test.

* tests/PB07-039__bad_mlw_with_flow_and_proof/test.out
Tag check is proved again.

Change-Id: I05925fd57158ac48d7db4ca3fd17f3d5a63ff033