TC02-027 update loc in output files in UG
In the UG, we sometimes only dispaly some lines of the output.
This is very fragile, and can be broken when the output changes.
Here I corrected the locations which were causing the doc build
to fail, plus an additional older one which we most likely missed
at some point. There might be other silent discrepencies that I
missed.
Also, either the test or the description in counterex_unprovable
should be changed, as the second counterexample disappeared.

Change-Id: Ie805c125ee9d1dfdf7f758b54185c86d1cd64207