UA11-004 fix for box in record aggregate
Record aggregates with boxes (<>) that represent components with no
default initialization are not allowed in SPARK. However, this violation
is detected in flow, not in marking, so flow itself must be prepared to
handle such boxes when they occur.

* src/flow/flow_utility.adb
(Merge): An empty Input parameter was only handled in one branch of the
if-statement. Now refactored, so it is always handled.

* testsuite/gnatprove/tests/UA11-004__flow_box/
Enable; add output.

Change-Id: I060926faffa2a6e76d2741e4f765b61ab415e1c0