TC02-027 Set location of precondition failures to the call
Before, a precondition failure during SPARK RAC carried the location
of the definition of the precondition, instead of the location of
the call.

* src/spark/spark_rac.adb:
(RAC_Call): Add a parameter N for the call node and set node of
precondition failures to it. Also, check preconditions before
executing the declarations.
* src/why/gnat2why-error_messages.adb:
(Handle_Result): Initialise the Small_Step_Res when the even
counterexample is empty
* testsuite/gnatprove/tests/*/test.out:
Rewrite test output with gained counterexamples

Change-Id: Iae66a685f64f70063029b25ac6505c934e31965f