UA19-012 Fix crash of build
With checking of pragma Predicate for presence of ghost entities added
in GNAT, the build of gnat2why was crashing due to a long-lasting bug
in the compilation of a predicate. Rewrite the predicate to avoid that
bug.

[changelog]
* src/flow/flow_types.ads (Global_Flow_Ids): Rewrite the predicate to
use quantification over cursors instead of quantification over elements.
As a result, we can now use the aspect form instead of a pragma.

Change-Id: Ie10484be6745f9a7e74c27fd3d25f04c63642f81