U726-022 Further update to aliasing in LRM
Anonymous access-to-constant "in" parameters are a complex corner
case for aliasing. The contents can be changed via an aliased
access-to-variable parameter so they are not (quite) "immutable".
However neither the referenced location nor the contents can be
changed via the access-to-constant parameter itself, so it can be
treated similarly to an immutable parameter in certain cases.

* docs/lrm/source/subprograms.rst
  Update verification rules 3 and 4.

Change-Id: I171735214dbe1076ca8b02c1ae35722cd12af6f8