UB04-014 assume function guard in EW_Pterm domain
When a function is called, we need to assume that this guard predicate
holds for its postcondition to be availlable. In the program domain, we
use the program function, which has the predicate in its postcondition.
In the term domain, we use an epsilon construct. In the Pterm domain, we
now add an assume statement.

* src/why/why-gen-expr.adb
(New_Function_Call): If the function has a guard predicate, assume it
 in the EW_Pterm domain.

* src/why/gnat2why-subprograms.adb
(Check_Ceiling_Protocol): Fix place where EW_Pterm was used instead of
 EW_Term.

* tests/UB04-014__guard_in_pterm
File test.

Change-Id: I28663e778b1a6f32058c628b955b49cf8b64d26c