UB22-046 Detect private part of descendants of private extensions
Flow now creates private parts for a type T when it derives
from a private type whose full view is not visible from the scope.
Previously, flow was unable to detect such private part, when the
private extension of the parent type was hidden under a
SPARK_Mode => Off.

* src/flow/flow_utility.adb
(Ancestor): Move function from declaration part of Unique_Components
to library-level.
(Flatten_Variable): Explore type ancestors to determine if an ancestor
is private and its full view is not visible from the scope. In such
cases, add a 'Private_Part to the result.

* testsuite/gnatprove/tests/UB22-046__flow_private_extension_init/
Flow now recognizes the private part of G.

Change-Id: Icbb42fe4aab7cf8373568546f11d48390fdde4d3