Remove deprecated "centered" Sphinx directive from SPARK RM
The ".. centered::" directive suffers from a number of issues. It is
described as:

   This directive creates a centered boldfaced line of text. Use it as
   follows:

   .. centered:: LICENSE AGREEMENT

   Deprecated since version 1.1: This presentation-only directive is a
   legacy from older versions. Use a rst-class directive instead and add
   an appropriate style.

It should itself produce a boldfaced text, yet we used it with explicit
strong emphasis (boldface). This was a workaround for a bug in Sphinx
LaTeX builder, which doesn't respect the documentation and doesn't
enable boldface; for HTML builder we were getting double emphasis, i.e.
"<strong><strong>...</strong></strong>".

There is little point in fixing a bug in a deprecated feature. The
suggested "rst-class" directive doesn't work with LaTeX either, but a
related "container" directive does work, so let's use it instead.

* docs/lrm/source/_static/lrm.css
Add a CSS style for "heading" in HTML output.
* docs/lrm/source/conf.py
Add a LaTeX style for "heading" in PDF output; enable custom CSS.

* docs/lrm/source/declarations-and-types.rst
* docs/lrm/source/exceptions.rst
* docs/lrm/source/generic-units.rst
* docs/lrm/source/names-and-expressions.rst
* docs/lrm/source/packages.rst
* docs/lrm/source/program-structure-and-compilation-issues.rst
* docs/lrm/source/representation-issues.rst
* docs/lrm/source/statements.rst
* docs/lrm/source/subprograms.rst
* docs/lrm/source/tasks-and-synchronization.rst
* docs/lrm/source/the-standard-library.rst
* docs/lrm/source/visibility-rules.rst
Convert deprecated "centered::" to new "container:: heading" with this
script:
  sed -i 's/^\.\. centered:: \*\*\([A-Za-z ]*\)\*\*/
            .. container:: heading\n\n   \1/' *.rst

no-tn-check

Change-Id: Iad7090ebfc6c14f12715ca1a70889494f13f038a