U915-006 set Ada_Node when transforming N_Aggregate_Extension
Set Ada_Node when transforming N_Aggregate_Extension, which appears in
the expanded wrapper for dispatching function with controlling result.
Setting it for New_Ada_Record_Access is necessary to avoid an assertion
failure; setting it for Generate_Associations_From_Ancestor is just
for consistency.

* src/why/gnat2why-expr.adb
(Transform_Expr): Set Ada_Node in branch for N_Aggregate_Extension
* testsuite/gnatprove/tests/PB07-039__bad_mlw_with_flow_and_proof/
Spurious checks are gone.
* testsuite/gnatprove/tests/U908-014__inherited_function_calls/test.out
Temporary regression; will be addressed in a follow-up commit.

Change-Id: Id2cc63a34a3e9b177dc83342a97e4751694b8117