Fix visibility issue for "**" operator
The "**" operator is predefined for every integer type and, therefore,
subject to the common visibility rules.  That's why a use (type) clause
is necessary here to make the code legal.

This will be diagnosed by an enhanced version of the compiler as such:

spark_atree-entities.adb:558:42:
  error: operator for type "Interfaces.Unsigned_64" is not directly
                                                                 visible
spark_atree-entities.adb:558:42:
  error: use clause would make operation legal

[changelog]
	* src/why/spark_atree-entities.adb (Max_Size_Of_Img_Attr): Add use
	type clause for Interfaces.Unsigned_64.

TN: UA26-051
Change-Id: I5b434044e5b4260185723eb81c7ce6e4539337ce