V104-014 Fix incorrect use of switches not specified for file
When using file-specific switches in a project file, specific switches
set for a given file might be applied to the analysis of subsequent
files. This does not result in unsoundness, but it might make the
analysis use switches that were not intended by the user, which may
impact verification results or time.

[changelog]
* src/gnatprove/configuration.adb (Read_Command_Line): Reset
file-specific switches before considering each file in turn.

Change-Id: I98c727a69ec1be7a12b9fced56d937dc634d0b8a