V106-025 crash on discrimiant dependant access to incomplete type
For all types, we generate a Dynamic_Invariant predicate stating what
it means for an object of the type to be valid. When there is an access
to an incomplete type inside the predicate of another incomplete type,
we cannot express directly that the value designated by the object if
any is valid, as the structure could be recursive so the predicate would
be infinite. So instead, we generate a separate predicate function for
the access to incomplete type. However, when the access to incomplete
type of the component is an Itype depending on the discriminant of the
enclosing type, the separate predicate function turns out to depend on
the discriminant which is not defined outside of the predicate of the
enclosing type.
We detect this case and generate the discriminant constraints directly
in the predicate of the enclosing type, and generate the separate
predicate for the underlaying unconstrained access to incomplete type
instead.
Along the way, fix 2 small other issues coming from incomplete types
with discriminant constraints.

* src/why/gnat2why-expr.adb
(Compute_Dynamic_Invariant): For Itypes with a discriminant constraint,
we handle the discriminant constraints before generating the separate
dynamic invariant predicate.

* src/why/gnat2why-types.adb
(Create_Dynamic_Invariant): Unrelated, give a name to the defining
 axiom for dynamic_invariants in case we generate several.

* src/why/why-gen-pointers.adb
(Declare_Conversion_Check_Function): Unrelated, use the type of the
 value designated by the root to get the discriminants of the root, in
 case the current pointer type and its root do not designate the same
 type.

* tests/V106-025__discr_dependent_incompl_type
New test.

Change-Id: I9ca3124c9f8c9dfaca8056e3dda18abc092d95d8