U915-006 pick SPARK_Pragma from the type for dispatching wrappers
The internally generated wrappers for dispatching functions with
controlling results inherit SPARK_Pragma from where they are frozen.
With this change we override that behaviour and pick the SPARK_Pragma
from where (the full view of) the derived type is declared.

Also, replace a special case for parameters of predicate functions
with a general recursion that works for all parameters, including
those of dispatching wrappers.

No test is affected by this change, but the change is consistent with
what we do for predicate functions.

* src/spark/spark_definition.adb
(SPARK_Pragma_Of_Entity): Handle wrappers for dispatching results.

Change-Id: Idc1e7bbbcf258e82653fecaf6c4a704c7709692f