V118-010 address a ??? comment in RAC counterexamples for characters
The uses of Name_Find in RAC were definitely wrong. The new code is
still dubious, because N_Character_Literal is a node and not entity,
but at least it is not junk.

* src/counterexamples/ce_rac.adb
(Character_Value): Fix use of Name_Find.
(Default_Value): Use Character_Value, which otherwise was unreferenced,
but GNAT failed to detect this because it gets confused by expression
functions.

Change-Id: I67de0558440e4c47e5905ff1647e3dbda93bc5f6