U607-011 Retype Get_Array_Attr to work on terms
As this function uses its parameter multiple times, it should be a
simple term, in general just a temporary constant introduced as proxy
for any expression (a prog or a term). Retype it to take and return a
term, and propagate this constraint to its callers, which requires
introducing temporaries in a few places.

The version taking a type needs to take the subdomain of type which is
relevant, EW_Term or EW_Pterm, as it can be used for dynamic bounds
of arrays which require translating the bound expression in the right
subdomain.

Remove apparently arbitrary uses of EW_Prog in various places, where no
check is expected, and a simple term can be used. Remove explicit
domains where the implicit one is obtained by caller resolution.

[changelog]
* src/why/gnat2why-expr.adb: Retype.
(Get_Item_From_Expr): Introduce a temporary for an array before each
call to obtain one of its bounds. It would be nice to factor this
temporary outside of the call, but that's not how immutable parameters
are handled for now, as they are supposed to be self-contained without
the need to wrap the call in a binding to introduce temporaries.
* src/why/why-gen-arrays.adb
(New_Length_Equality): Retype arguments to take terms.
(Get_Array_Attr): Take and return terms.
(New_Array_Range_Expr): Take term for the index/range arguments.
* src/why/why-gen-arrays.ads: Same.
* src/why/why-gen-expr.adb
(Insert_Array_Index_Check, Insert_Length_Check): Take terms.
(Do_Index_Check): Take term for array parameter (not for index).
(New_Dynamic_Property): Take term.
* src/why/why-gen-expr.ads: Same.
* src/why/why-gen-pointers.adb: Retype.
* src/why/why-sinfo.ads: Fix documentation on EW_Pterm vs EW_Term.

Change-Id: Ib9aa9989eb67b007a4565670cd28dc2f064ed57f