V215-001 remove duplicated checks on the prefix of updates
When a value of a subcomponent of an object is updated from its
previous value, it is not necessary to perform the checks that ensure
that the subcomponent is a valid part of the object twice (once during
the computation of the previous value and once during the update).

* src/why/gnat2why-expr.adb
(New_Assignment): No_Check boolean flag is replace by a three-valued
 Do_Check flag which can be All_Checks, No_Checks, or Only_Vars. The
 third value is used to express that checks which are preserved by the
 form of the prefix (index checks, discriminant checks...) do not need
 to be redone.
(One_Level_Update, Shift_Rvalue): Add a Check_Prefix flag to disable
 the checks that are preserved by the form of the prefix.
(Compute_Store): Call New_Assignment with Do_Check set to Only_Vars.
(Transform_Assignment_Statement): Call New_Assignment with Do_Check set
 to Only_Vars if the assigned value contains an occurrence of the
 target name @.

* tests/OC01-011__justify_aliasing/test.out
* tests/P113-023__alias_issue/test.out
* tests/P630-025__loopinv/test.out
* tests/P630-025__task_init/test.out
* tests/P630-037__keccak/test.out
* tests/P915-005__flow_array_initialization_via_entry/test.out
* tests/P915-005__flow_array_initialization_via_proc/test.out
* tests/Q502-001__target_name/test.out
* tests/RC17-014__init_by_proof/test.out
* tests/S117-008__pointer/test.out
* tests/T221-052__flow_alias_slice/test.out
* tests/adacore_u/test.out
Remove duplicated checks.

* tests/ghc_sort
Update session test.

Change-Id: I3e29d9d58ebba4263cb7423e6eac6e65b088ed7e