V209-039, V420-041 Prototype for excluding lines from analysis
This is a crude prototype to experiment with the idea of excluding
lines of code from analysis, as a way to check whether an uncovered
line of code during testing is relevant for the proof.

A new switch --exclude-line=file:line is added to specify the line
to exclude. This switch is intentionally not documented as part of
the help message or User's Guide, so that it can be removed later.

[changelog]
* src/common/gnat2why_opts.ads (Exclude_Line_Name): New name,
* src/flow/flow_error_messages.adb (Is_Excluded_Line): New query.
* src/flow/flow_error_messages.ads: Same.
* src/gnatprove/configuration.adb (Parse_Switches): Add new switch.
* src/gnatprove/configuration.ads (CL_Switches): Add new field.
* src/gnatprove/gnat2why_opts-writing.adb (To_JSON): Write new field.
* src/spark/gnat2why_opts-reading.adb (Read_File_Specific_Info): Read
new field.
* src/spark/gnat2why_opts-reading.ads (Exclude_Line): Add new variable.
* src/spark/spark_rewrite.adb (Rewrite_Node): Delete lines of code
corresponding to statements or declarations whose sloc matches the one
passed in argument to --exclude-line.

Change-Id: I62a3ea48c4c661bd66034395f4ecc0a5bfc83933