UA27-029 implement more accurate array equality check
The length of the the arrays and their values and other fields are now
properly taken into account when determining if two Value_Types of array
kind are equal, i.e. contain the same elements in the same order. A
RAC_Stuck exception is raised if computing the length of the arrays is
impossible due to missing First or Last information. A constraint error
is raise if the length of the array cannot be stored in a Natural.

The comparison uses a for loop checking each element. Since the arrays
use sparse notation, this could be improved by only checking the
elements which we know can differ between the two.

* src/counterexamples/ce_rac.adb
  ("=" on Value_Type): Fix the array case to do proper comparison
  ("=" on Big_Integer_To_Value_Maps.Map): Remove now unused function

* src/counterexamples/ce_values.adb
  src/counterexamples/ce_values.ads
  (Get_Array_Elt): Implement function to get element at index in array
  (Get_Array_Length): Implement function to return length of Value_Type

* testsuite/gnatprove/tests/UA27-029__counterex_arr_equal/palindrom.adb
* testsuite/gnatprove/tests/UA27-029__counterex_arr_equal/test.out
  Add test with empty string equality precondition

* tests/riposte__arrays_in_records/test.out
* tests/riposte__arrays_of_records/test.out
* tests/riposte__complex_arrays/test.out
* tests/riposte__simple_arrays/test.out
  Update expected outputs to account for new behaviour

Change-Id: I57426b6546aa79445141fc23e46126796516de84