V209-031 use Why equality for private types with mode Off when possible
In general, if the full view of a private type is annotated with
SPARK_Mode Off, then we cannot assume anything about its predefined
equality, in particular, we cannot translate it using the Why3 equality
symbol (standing for the mathematical equality).
However, there are some cases where it is safe to use the mathematical
equality for the type (for discrete types, or record or constrained
arrays of acceptable types). We now peek into the full view of private
types with SPARK_Mode Off to detect this case.

* src/spark/spark_util-types.adb
(Use_Real_Eq_For_Private_Type): Check if the predefined equality on a
 private type opaque for SPARK can be represented using Why3 equality.

* src/why/why-gen-records.adb
(Declare_Rep_Record_Type): Use Why3 equality as the definition of the
 predefined equality on private types if possible.

* tests/M221-017__private_with_discr
Add user defined equality to retain previous behavior.

* tests/Q515-026__sk
Update test output.

* tests/SC03-021__binary_search_trees
* tests/T727-043__mergesort
Update session test.

* tests/V209-031__private_eq
New test.

Change-Id: I66143a3e398477613eaf73829af5bd94963be36a