V302-028 give a definition to predefined equality on hardcoded units
We now give a definition to the predefined equality operator bool_eq
on hardcoded types. It should theoretically not be necessary as the
primitive equality should always be used on private types whose full
view is a record type. As giving such a definition is harmless, it
seems a good idea as a hopefully temporary work-around.

* src/why/why-gen-hardcoded.ads
(Hardcoded_Equality_Symbol): Return the equality symbol for an
 hardcoded type.

* src/why/why-gen-records.adb
(Declare_Equality_Function): Give a definition to the predefined
 equality on hardcoded types.

* tests/V302-028__big_int_eq
File test.

Change-Id: I90f238b709594ecc3596e7063ecfbe4a323481c6