Spurious error about a volatile actual is gone
The affected test includes an object whose type is a subtype of a
private type; the completion of this private type has aspects
Volatile and Async_Readers, so it is not "volatile for reading".

Now this object is accepted when used as an actual parameter.

* testsuite/gnatprove/tests/V225-029__volatile_private/test.out
Spurious error in a minimal reproducer is gone.

Change-Id: I95bcb1ea7d91c55b13551d2d87736dafafd4d11c
TN: V225-029