V310-001 take into account prover timeout for memcached
Prover timeout is not given via the command line, instead it is
imposed via kernel resource limits. This means that memcached
doesn't "see" the time limit and might recover a proof result that
used a different (e.g. higher) time limit than the current run. This
may lead to a run with cache enabled proving more than a run without
(though not strictly unsound as the proved VCs were really proved in
the past).

We solve this by adding an extra argument to the
spark_memcached_wrapper, which indicates extra information to be
used for hashing purposes. In the case of provers, we pass the
timeout, in the case of gnatwhy3 we pass a dummy string "salt".

* configuration.adb
(Compute_Why3_Args): add dummy salt to invocation of
  spark_memcached_wrapper for gnatwhy2
* gnatprove.adb
(Generate_Why3_Conf): add timeout as salt to invocation of
  spark_memcached_wrapper for provers
* spark_memcached_wrapper.adb
(Hash_CommandLine, Init_Client, Main): adapt indices accessing
  arguments
(Compute_Key): hash salt first

Change-Id: I3ab8b02f674fdb52413ca0b0c6d41e61836e11c1