V217-041 use partial view to query the user equality when present
* src/spark/spark_definition.adb
Fill the partial view map in Mark_Type_Entity instead of when marking
the declaration so that it is done on pulled entities.

* src/spark/spark_util-types.adb
(Get_User_Defined_Eq): Use the partial view when present.

* tests/S729-021__predicates/test.out
* tests/SC31-001__vector/test.out
The partial view map is used in Retyps to recognize the subtype
introduced for formal parameters of generic units. This removes useless
conversions and trivial checks in this case.

* tests/V217-041__user_eq
File test.

Change-Id: Ie2d1791316619a9f07663426d54a7d6c7a8ff02e