V310-013 handle usage of 'Old attribute for non-entity expressions
Until now, the usage of 'Old was only possible on entities. Now it is
possible to check for the old value of complex expressions like
F(X)'Old.

* src/counterexamples/ce_rac.adb
(Scopes): Change the content of scopes and adapt every usage of scopes
to this change.
Instead of storing a Binding per Entity_Id, store two maps:
    - One linking entity nodes to a Bindings
    - The other linking a node to an attributes map
The first one works the same way scopes did before this change. The
second one allows for the use of special attributes that can be used not
only on entity nodes but on every type of node.
This change only makes use of it for the 'Old attribute but it should
also be used for the handling of the 'Loop_Entry attribute. For
simplicity, these attributes' values will always be stored in the
Other_Attrs map even if their prefix is an entity.
(RAC_Call): Adapt the evaluation and storage of old values to account
for the new feature.
(RAC_Expr): Fetch old values correctly in the new map.
(Iterate_Loop_Param_Spec): Store iteration variable value in the current
scope instead of creating a new one.

* tests/K506-012__old/test.out
* tests/U301-038__misplaced_old/test.out
* tests/M614-039__simple/test.out
* tests/N616-037__vscomp2014/test.out
Update outputs due to improvements upon usage of 'Old attribute.

* tests/V310-013__counterex_old_on_expr/
Rename from V310-013__counterex_old_no_entity and update output to
account for new case added to the test.

* tests/V310-013__rac/main.adb
Add test based on V310-013__counterx_old_expr to test small step RAC
execution.

Change-Id: If1c3e93a77bd4c91bb05cb98f0b48b62f20e5928