V328-020 handle case expressions and ranges/subtypes in case statements
* src/counterexamples/ce_rac.adb
(RAC_Statement) : Support checking if the expression is in a given
range, including subtype ranges.
(RAC_Expr) : Support all case expressions.
(Check_Range) : Implement function to check if a Value_Type falls in the
range defined by a N_Range node.
(Check_Subtype) : Implement function to check if a Value_Type is of a
given subtype, should it be defined with a range, an expression or a
static predicate.
(Match_Case_Alternative) : Extract case alternative matching algorithm
from RAC_Statement in order to use it both for case statements and case
expressions.

* tests/V328-020__counterex_case_stmt/test.out
* tests/V328-020__counterex_case_expr/test.out
* tests/P216-021__inlining_recursive/test.out
* tests/KB10-018__cursor_location/test.out
* tests/R508-015__ce_branch/test.out
* tests/Q915-037__counterex/test.out
* tests/P804-001__driver/test.out
* tests/OC10-038__tetris/test.out
Update outputs

* tests/N523-032__hello/test.opt
Expect test to fail. The execution has previously halted at the
evaluation of case expressions which was unsupported. The execution can
now go further but crashes when an negative length string is given to
the concatenation operation.

Change-Id: I49d1564b9e57937c777767228c96516aac2d7f88