V328-020 fix empty string check in concatenation
* src/counterexamples/ce_rac.adb
(RAC_Binary_Op) : To check if the strings to concatenate were empty
their lengths were computed and compared to 0. However a string
starting at index 0 and finishing at -1 with a length of -1 is valid
and empty but wasn't detected as such. This lead to a crash when
trying to insert a character at a bad index. The comparison now
checks if the length is less than or equal to 0.

* tests/N114-033__null_concat/test.out
Update test output

* tests/N523-032__hello/test.opt
Remove file to expect failure since the test should not crash upon
concatenation of a negative length string anymore.

Change-Id: I5bb0cdd4c35df86bb21d9b4105d4c9d37962a206