V318-027 fix violation of tampering checks in RAC
When "lazily initializing" global constant without variable input in RAC
in routine Find_Binding, we were creating a cursor to the top-most
environment and then modified this environment while recursively doing
RAC on the constant's initial expression in the call to Get_Value.

This is now fixed by not creating a cursor in Find_Binding and
hardcoding the use of the top-most environment in Init_Global.

* src/counterexamples/ce_rac.adb
(Init_Global): Remove Scope parameter; instead, hardcode the use of the
top-most environment.
(Find_Binding): Adapt client.
(Init_Global_Scope): Adapt client (here we used the inner-most
environment but actually in this routine the environment stack has only
one element, so First and Last cursors are equivalent).

Change-Id: I0c4d61ae75c795da704757fa26db3c426add0327