V209-030 support by and so connective inside assertions
In Why3, 'by' and 'so' are boolean connectives which can be used to
help the proof by supplying cut assertions. If A and B are boolean
expressions, (A by B) and (B so A) are logically equivalent to A, but B
is used to help to prove A. Basically, when an assertion (A by B) is
encountered, the tool will attempt to prove B and the side-condition
(B -> A) and then assume A. It is similar for So: an assertion (B so A)
will require the proof of B and (B -> A), and will assume (B /\ A).
In SPARK, By and So are defined as functions in the spark lemma
library. They are hardcoded in GNATprove, so that their translation
works as described above.

* docs/ug/en/source/spark_libraries.rst
Document By and So in the UG.

* include/spark-cut_operations.ads
(By, So): New boolean binary operations. For execution they are handled
 as a conjunction.

* src/common/vc_kinds.ads
(VC_Kind): New kinds for assertion premises and steps.

* src/gnatprove/spark_report.adb
* src/why/gnat2why-error_messages.adb
Support new VC kinds.

* src/spark/spark_definition.adb
(Mark): Small unrelated fix, mark filter in loop parameter
 specification if any.
(Mark_Call): Check that cut operations occurs in a supported context.

* src/spark/spark_util-hardcoded.adb
The functions By and So from SPARK.Cut_Operations are hardcoded.

* src/spark/spark_util.ads
(Contains_Cut_Operations): Check whether the expression contains
 calls to cut operations.

* src/why/gnat2why-expr.adb
(Generate_Case_Expression): Simplify the profile to remove the expected
 type as it is not needed.
(Generate_Expr_With_Actions): Generic operation to handle expressions
 with actions.
(Transform_Expr_With_Actions): Instantiate Generate_Expr_With_Actions.
(Transform_Expr_With_Cutpoints): Transforms a boolean expression
 containing occurrences of By and So.
(Transform_Pragma_Check): Special case assertions containing
 occurrences of By and So.
(Transform_Statement_Or_Declaration_In_List): Special case
 assert_and_cut containing occurrences of By and So.

* src/why/why-gen-hardcoded.adb
Cut operations should never be handled here.

* testsuite/gnatprove/lib/python/test_support.py
Add new check markers.

* tests/V209-030__cut_operations
* tests/V209-030__cut_operation_context
New tests.

Change-Id: Idee42b5a44e84bc082fe17f32cd66ee5b1cbec84